Moved an inner class to top level.
