+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/VariantBackedConfigurationMetadata.java
@@ -1,0 +1,123 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.component.external.model;
+
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
+import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
+import org.gradle.api.internal.artifacts.DefaultModuleVersionSelector;
+import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.ModuleExclusion;
+import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.ModuleExclusions;
+import org.gradle.api.internal.attributes.ImmutableAttributes;
+import org.gradle.internal.Describables;
+import org.gradle.internal.DisplayName;
+import org.gradle.internal.component.model.ComponentArtifactMetadata;
+import org.gradle.internal.component.model.ConfigurationMetadata;
+import org.gradle.internal.component.model.DependencyMetadata;
+import org.gradle.internal.component.model.GradleDependencyMetadata;
+import org.gradle.internal.component.model.IvyArtifactName;
+import org.gradle.internal.component.model.VariantMetadata;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Set;
+
+class VariantBackedConfigurationMetadata implements ConfigurationMetadata {
+    private final ModuleComponentIdentifier componentId;
+    private final ComponentVariant variant;
+
+    VariantBackedConfigurationMetadata(ModuleComponentIdentifier componentId, ComponentVariant variant) {
+        this.componentId = componentId;
+        this.variant = variant;
+    }
+
+    @Override
+    public String getName() {
+        return variant.getName();
+    }
+
+    @Override
+    public Collection<String> getHierarchy() {
+        return ImmutableList.of(variant.getName());
+    }
+
+    @Override
+    public String toString() {
+        return asDescribable().getDisplayName();
+    }
+
+    @Override
+    public DisplayName asDescribable() {
+        return Describables.of(componentId, "variant", variant.getName());
+    }
+
+    @Override
+    public ImmutableAttributes getAttributes() {
+        return variant.getAttributes().asImmutable();
+    }
+
+    @Override
+    public Set<? extends VariantMetadata> getVariants() {
+        return ImmutableSet.of(variant);
+    }
+
+    @Override
+    public boolean isCanBeConsumed() {
+        return true;
+    }
+
+    @Override
+    public boolean isCanBeResolved() {
+        return false;
+    }
+
+    @Override
+    public boolean isTransitive() {
+        return true;
+    }
+
+    @Override
+    public boolean isVisible() {
+        return true;
+    }
+
+    @Override
+    public ModuleExclusion getExclusions(ModuleExclusions moduleExclusions) {
+        return ModuleExclusions.excludeNone();
+    }
+
+    @Override
+    public ComponentArtifactMetadata artifact(IvyArtifactName artifact) {
+        return new DefaultModuleComponentArtifactMetadata(componentId, artifact);
+    }
+
+    @Override
+    public Set<? extends ComponentArtifactMetadata> getArtifacts() {
+        return ImmutableSet.of();
+    }
+
+    @Override
+    public List<? extends DependencyMetadata> getDependencies() {
+        // TODO - Should calculate this value once
+        List<DependencyMetadata> dependencies = new ArrayList<DependencyMetadata>(variant.getDependencies().size());
+        for (ComponentVariant.Dependency dependency : variant.getDependencies()) {
+            dependencies.add(new GradleDependencyMetadata(DefaultModuleVersionSelector.newSelector(dependency.getGroup(), dependency.getModule(), dependency.getVersion())));
+        }
+        return dependencies;
+    }
+}
